Rename java package 'dummy' to 'sample' 'Sample' is more explicit than 'dummy'
