--- src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/dummy/feature/package-info.java
@@ -1,2 +1,0 @@
-@tech.jhipster.lite.BusinessContext
-package tech.jhipster.lite.generator.server.springboot.mvc.dummy.feature;
