+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/sample/feature/application/DummyApplicationService.java
@@ -1,0 +1,20 @@
+package tech.jhipster.lite.generator.server.springboot.mvc.sample.feature.application;
+
+import org.springframework.stereotype.Service;
+import tech.jhipster.lite.generator.server.springboot.mvc.sample.feature.domain.DummyFeatureModuleFactory;
+import tech.jhipster.lite.module.domain.JHipsterModule;
+import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+
+@Service
+public class DummyApplicationService {
+
+  private final DummyFeatureModuleFactory factory;
+
+  public DummyApplicationService() {
+    factory = new DummyFeatureModuleFactory();
+  }
+
+  public JHipsterModule buildModule(JHipsterModuleProperties properties) {
+    return factory.buildModule(properties);
+  }
+}
