+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/sample/flyway/application/DummyFlywayApplicationService.java
@@ -1,0 +1,24 @@
+package tech.jhipster.lite.generator.server.springboot.mvc.sample.flyway.application;
+
+import org.springframework.stereotype.Service;
+import tech.jhipster.lite.generator.server.springboot.mvc.sample.flyway.domain.DummyFlywayModuleFactory;
+import tech.jhipster.lite.module.domain.JHipsterModule;
+import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+
+@Service
+public class DummyFlywayApplicationService {
+
+  private final DummyFlywayModuleFactory factory;
+
+  public DummyFlywayApplicationService() {
+    factory = new DummyFlywayModuleFactory();
+  }
+
+  public JHipsterModule buildPostgresqlModule(JHipsterModuleProperties properties) {
+    return factory.buildPostgresqlModule(properties);
+  }
+
+  public JHipsterModule buildNotPostgresqlModule(JHipsterModuleProperties properties) {
+    return factory.buildNotPostgresqlModule(properties);
+  }
+}
