+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/sample/liquibase/infrastructure/primary/DummyLiquibaseModuleConfiguration.java
@@ -1,0 +1,27 @@
+package tech.jhipster.lite.generator.server.springboot.mvc.sample.liquibase.infrastructure.primary;
+
+import static tech.jhipster.lite.generator.slug.domain.JHLiteFeatureSlug.*;
+import static tech.jhipster.lite.generator.slug.domain.JHLiteModuleSlug.*;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import tech.jhipster.lite.generator.server.springboot.mvc.sample.liquibase.application.DummyLiquibaseApplicationService;
+import tech.jhipster.lite.module.domain.resource.JHipsterModuleOrganization;
+import tech.jhipster.lite.module.domain.resource.JHipsterModuleResource;
+
+@Configuration
+class DummyLiquibaseModuleConfiguration {
+
+  @Bean
+  JHipsterModuleResource dummyLiquibaseModule(DummyLiquibaseApplicationService dummyLiquibase) {
+    return JHipsterModuleResource.builder()
+      .slug(DUMMY_LIQUIBASE_CHANGELOG)
+      .withoutProperties()
+      .apiDoc("Spring Boot - MVC", "Add liquibase changelog for dummy feature")
+      .organization(
+        JHipsterModuleOrganization.builder().feature(DUMMY_SCHEMA).addDependency(LIQUIBASE).addDependency(DUMMY_FEATURE).build()
+      )
+      .tags("server")
+      .factory(dummyLiquibase::buildModule);
+  }
+}
