+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/sample/mongopersistence/domain/DummyMongoDBPersistenceModuleFactory.java
@@ -1,0 +1,42 @@
+package tech.jhipster.lite.generator.server.springboot.mvc.sample.mongopersistence.domain;
+
+import static tech.jhipster.lite.module.domain.JHipsterModule.*;
+
+import tech.jhipster.lite.module.domain.JHipsterModule;
+import tech.jhipster.lite.module.domain.file.JHipsterSource;
+import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
+import tech.jhipster.lite.shared.error.domain.Assert;
+
+public class DummyMongoDBPersistenceModuleFactory {
+
+  private static final JHipsterSource SOURCE = from("server/springboot/mvc/dummy/mongodbpersistence");
+
+  private static final String SECONDARY = "infrastructure/secondary";
+  private static final String SECONDARY_DESTINATION = "dummy/" + SECONDARY;
+
+  public JHipsterModule buildModule(JHipsterModuleProperties properties) {
+    Assert.notNull("properties", properties);
+
+    String packagePath = properties.packagePath();
+
+    //@formatter:off
+    return moduleBuilder(properties)
+      .files()
+        .batch(SOURCE.append("main").append(SECONDARY), toSrcMainJava().append(packagePath).append(SECONDARY_DESTINATION))
+          .addTemplate("BeerDocument.java")
+          .addTemplate("BeersCollectionChangeUnit.java")
+          .addTemplate("MongoDBBeersRepository.java")
+          .addTemplate("SpringDataBeersRepository.java")
+          .and()
+        .batch(SOURCE.append("test").append(SECONDARY), toSrcTestJava().append(packagePath).append(SECONDARY_DESTINATION))
+          .addTemplate("BeerDocumentTest.java")
+          .addTemplate("MongoDBBeersRepositoryIT.java")
+          .addTemplate("MongoDBBeersResetter.java")
+          .and()
+        .delete(path("src/main/java").append(packagePath).append(SECONDARY_DESTINATION).append("InMemoryBeersRepository.java"))
+        .delete(path("src/test/java").append(packagePath).append(SECONDARY_DESTINATION).append("InMemoryBeersResetter.java"))
+        .and()
+      .build();
+    //@formatter:on
+  }
+}
